refactor gt generation in eval
